backend/pyproject.toml
[project]
name = "kratos-backend"
version = "0.1.0"
description = "FastAPI backend for verifying KratosComply reports"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.111.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.7.0",
  "sqlalchemy>=2.0.0",
  "python-dotenv>=1.0.0",
  "httpx>=0.27.0",
  "psycopg2-binary>=2.9.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "httpx>=0.27.0"
]
